Create Context for Header Visibility
I created a context for the header visibility and created functionallity
to show the size that has been selected for the chosen item on the item page by
blacking out the boarder of the selected size that has been clicked. I also made it when an item
is added to to the cart the header will be set to visible, and if a size is not selected
the add to cart button will be disabled.